testing

+++ backend/backend_setup/waddlewaitMenu/models.py
@@ -2,7 +2,7 @@ from django.db import models
 
 class Category(models.Model):
     name = models.CharField(max_length=100)
-    display_order = models.DecimalField(max_digits=8, default = 99999) #thus new items are shown last
+    display_order = models.IntegerField(default = 99999) #thus new items are shown last
     class Meta:
         ordering = ['display_order']
 
@@ -16,7 +16,7 @@ class MenuItem(models.Model):
     price = models.DecimalField(max_digits=8, decimal_places=2)
     image = models.ImageField(upload_to='menu_images/', default='menu_images/default_image.jpg')
     category = models.ForeignKey(Category, on_delete=models.CASCADE, related_name='menu_items')
-    display_order = models.DecimalField(max_digits=8, default = 99999) #thus new items are shown last
+    display_order = models.IntegerField(default = 99999) #thus new items are shown last
     
     class Meta:
         ordering = ['display_order']
